Style Updates, Create Submissions, Add Task Specs
Updating the task index and task show page with styles more consistant
to what it will end up looking like.

Adding the submissions model so when a user completes a task they can
submit it. The submission model handles any text input they need to fill
out to complete the task and will allow file uploads to be attached to
it to provide better in app support for completing tests and passing
assets back and forth.

Fixed some specs that were now failing due to changes in production and
added a feature spec around creating a task as a company admin.