Mixpanel Tracking Start
This is the first implementation of adding pageview tracking to
mixpanel. It is turned off in development and local modes to save some
requests. A meta tag was used to turn it on and off to quickly be able
to access the status from JS instead of using ERB to conditionally
render partials.