Reorganizing Assets
I decided to go with a more complicated app like layout. In order to
support this I want to keep the existing layout but moved it under the
public.html layout file. This is for all public pages on the site.

The application files got moved under the application layout. This is
for the core application once a user signs up. The assets were broken up
to mirror this division. Top Level Directories contain a 'shared',
'public' and 'application' directory. Shared is the base bourbon and
layouts while the other two split the assets into public or application
creating a clean seperation.