Moving to User Based Task System
I decided to move away from training and department type training tasks
and reutilize the code for letting any user create tasks for any other
user. The intent is to allow people anywhere in the company to gain
benefit from the platform.

Created "Thank You" system in which users can thank other users in the
company. This is going to be a core part of the system. Users post
random things and then can thank the people that help them.